+++ recipes/dev_tools.rb
@@ -53,10 +53,6 @@ end
 
 execute '/usr/sbin/php5enmod xdebug'
 
-# Configure xdebug settings
-node.default['php']['xdebug']['idekey']          = 'PHPSTORM'
-node.default['php']['xdebug']['ide_server_name'] = node['hostname']
-
 # Install the xdebug CLI wrapper script
 template '/usr/local/bin/xdebug' do
   source    'xdebug.erb'